pyproject.toml
[project]
name = "nationaid-number"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-api-key>=3.1.0",
    "pytest>=8.4.2",
]
